pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-template"
version = "0.1.0"
description = "A template project for building AI agent applications."
authors = [{name = "bartosz.golebiowski"}]
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "openai>=1.13.3",
]

[tool.setuptools.packages.find]
where = ["."]
